added scroll hotkeys to text novel view too
