+++ src/components/collection-explorer/EpubNovelView.tsx
@@ -1,4 +1,4 @@
-import { FC, HTMLAttributes, useCallback, useEffect, useRef, useState } from 'react';
+import { forwardRef, HTMLAttributes, useEffect, useRef, useState } from 'react';
 import { useTheme } from '@/hooks/use-theme';
 import epub, { Rendition } from 'epubjs';
 import { convertFileSrc } from '@tauri-apps/api/tauri';
@@ -6,8 +6,6 @@ import { themeColors } from '@/styles/theme-colors';
 import { twMerge } from 'tailwind-merge';
 import { NovelAsset } from '@/types/collection';
 import { Theme } from '@/types/theme';
-import { useAnimateScroll, useKeyboardEvent } from '@/hooks';
-import { checkTextfieldFocused } from '@/utils/is-textfield-focused';
 
 const getRenditionTheme = (theme: Theme) => ({
   body: {
@@ -29,116 +27,51 @@ export interface EpubNovelDocumentProps extends HTMLAttributes<HTMLDivElement> {
   asset: NovelAsset;
 }
 
-export const EpubNovelDocument: FC<EpubNovelDocumentProps> = ({ asset, ...divProps }) => {
-  const { theme } = useTheme();
-  const [rendition, setRendition] = useState<Rendition | null>(null);
-  const viewWrapperRef = useRef<HTMLDivElement | null>(null);
-  const scrollContainerRef = useRef<Element | null>(null);
-  const animateScroll = useAnimateScroll(scrollContainerRef);
+export const EpubNovelDocument = forwardRef<Element | null, EpubNovelDocumentProps>(
+  ({ asset, ...divProps }, ref) => {
+    const { theme } = useTheme();
+    const [rendition, setRendition] = useState<Rendition | null>(null);
+    const viewWrapperRef = useRef<HTMLDivElement>(null);
 
-  useEffect(() => {
-    const viewWrapper = viewWrapperRef.current;
-    if (!viewWrapper) return;
+    useEffect(() => {
+      const viewWrapper = viewWrapperRef.current;
+      if (!viewWrapper) return;
 
-    const abortController = new AbortController();
-    const book = epub(convertFileSrc(asset.path));
-    const rendition = book.renderTo(viewWrapper, { manager: 'continuous', flow: 'scrolled' });
-    setRendition(rendition);
+      const abortController = new AbortController();
+      const book = epub(convertFileSrc(asset.path));
+      const rendition = book.renderTo(viewWrapper, { manager: 'continuous', flow: 'scrolled' });
+      setRendition(rendition);
 
-    return () => {
-      abortController.abort();
-      book.destroy();
-    };
-  }, [asset]);
+      return () => {
+        abortController.abort();
+        book.destroy();
+      };
+    }, [asset]);
 
-  useEffect(() => {
-    rendition?.themes.default(getRenditionTheme(theme));
-  }, [rendition, theme]);
+    useEffect(() => {
+      rendition?.themes.default(getRenditionTheme(theme));
+    }, [rendition, theme]);
 
-  useEffect(() => {
-    rendition?.display(0).then(() => {
-      scrollContainerRef.current =
-        viewWrapperRef.current?.getElementsByClassName('epub-container')[0] ?? null;
-    });
-  }, [rendition]);
+    useEffect(() => {
+      rendition?.display(0).then(() => {
+        if (!ref) return;
 
-  const scrollBy = useCallback(
-    (offset: number) => {
-      const scrollContainerElement = scrollContainerRef.current;
-      if (!scrollContainerElement) return;
-
-      const currentScrollTop = scrollContainerElement.scrollTop;
-      animateScroll.start({
-        from: { y: scrollContainerRef.current?.scrollTop },
-        y: currentScrollTop + offset,
-        reset: true,
+        const scrollContainer =
+          viewWrapperRef.current?.getElementsByClassName('epub-container')[0] ?? null;
+        if (typeof ref === 'function') ref(scrollContainer);
+        else ref.current = scrollContainer;
       });
-    },
-    [scrollContainerRef, animateScroll],
-  );
+    }, [rendition, ref]);
 
-  const scrollTo = useCallback(
-    (offset: number) => {
-      const scrollHeight = scrollContainerRef.current?.scrollHeight;
-
-      if (offset >= 0) animateScroll.start({ y: offset });
-      else if (scrollHeight) animateScroll.start({ y: scrollHeight + offset + 1 });
-    },
-    [animateScroll],
-  );
-
-  useKeyboardEvent(
-    'keydown',
-    ['PageUp', 'ArrowLeft', 'KeyA'],
-    (e) => {
-      if (checkTextfieldFocused() || !rendition) return;
-      e.preventDefault();
-      scrollBy(e.code === 'PageUp' ? -300 : -150);
-    },
-    [rendition],
-  );
-
-  useKeyboardEvent(
-    'keydown',
-    ['PageDown', 'ArrowRight', 'KeyD'],
-    (e) => {
-      if (checkTextfieldFocused() || !rendition) return;
-      e.preventDefault();
-      scrollBy(e.code === 'PageDown' ? 300 : 150);
-    },
-    [rendition],
-  );
-
-  useKeyboardEvent(
-    'keydown',
-    ['Home'],
-    (e) => {
-      if (checkTextfieldFocused() || !rendition) return;
-      e.preventDefault();
-      scrollTo(0);
-    },
-    [scrollTo],
-  );
-
-  useKeyboardEvent(
-    'keydown',
-    ['End'],
-    (e) => {
-      if (checkTextfieldFocused() || !rendition) return;
-      e.preventDefault();
-      scrollTo(-1);
-    },
-    [scrollTo],
-  );
-
-  return (
-    <div
-      {...divProps}
-      ref={viewWrapperRef}
-      className={twMerge(
-        divProps.className,
-        '[&_.epub-container]:!size-full [&_.epub-view]:!w-full [&_iframe]:!w-full',
-      )}
-    />
-  );
-};
+    return (
+      <div
+        {...divProps}
+        ref={viewWrapperRef}
+        className={twMerge(
+          divProps.className,
+          '[&_.epub-container]:!size-full [&_.epub-view]:!w-full [&_iframe]:!w-full',
+        )}
+      />
+    );
+  },
+);
