+++ src/components/collection-explorer/TextNovelDocument.tsx
@@ -1,4 +1,4 @@
-import { FC, HTMLAttributes, useEffect, useState } from 'react';
+import { ForwardedRef, forwardRef, HTMLAttributes, useEffect, useState } from 'react';
 import { convertFileSrc } from '@tauri-apps/api/tauri';
 import { twMerge } from 'tailwind-merge';
 import { NovelAsset } from '@/types/collection';
@@ -7,29 +7,32 @@ export interface TextNovelDocumentProps extends HTMLAttributes<HTMLPreElement> {
   asset: NovelAsset;
 }
 
-export const TextNovelDocument: FC<TextNovelDocumentProps> = ({ asset, ...preProps }) => {
-  const [text, setText] = useState<string | null>(null);
+export const TextNovelDocument = forwardRef<Element, TextNovelDocumentProps>(
+  ({ asset, ...preProps }, ref) => {
+    const [text, setText] = useState<string | null>(null);
 
-  useEffect(() => {
-    const abortController = new AbortController();
+    useEffect(() => {
+      const abortController = new AbortController();
 
-    fetch(convertFileSrc(asset.path), { signal: abortController.signal })
-      .then((response) => response.text())
-      .then((text) => setText(text))
-      .catch(() => {
-        if (abortController.signal.aborted) return;
-        setText(null);
-      });
+      fetch(convertFileSrc(asset.path), { signal: abortController.signal })
+        .then((response) => response.text())
+        .then((text) => setText(text))
+        .catch(() => {
+          if (abortController.signal.aborted) return;
+          setText(null);
+        });
 
-    return () => abortController.abort();
-  }, [asset]);
+      return () => abortController.abort();
+    }, [asset]);
 
-  return (
-    <pre
-      {...preProps}
-      className={twMerge(preProps.className, 'overflow-y-scroll px-[10%] py-8 font-[inherit]')}
-    >
-      {text}
-    </pre>
-  );
-};
+    return (
+      <pre
+        ref={ref as ForwardedRef<HTMLPreElement>}
+        {...preProps}
+        className={twMerge(preProps.className, 'overflow-y-scroll px-[10%] py-8 font-[inherit]')}
+      >
+        {text}
+      </pre>
+    );
+  },
+);
